finish eslint config
