+++ craco.config.js
@@ -5,14 +5,20 @@ module.exports = {
       plugins: [
         require('tailwindcss'),
         require('autoprefixer'),
-        require('babel-plugin-module-resolver', {
-          "root": ["."],
-          "extensions": [".js", ".jsx"],
-          "alias": {
-            "@components": "./src/components"
-          }
-        })
+
       ],
     },
   },
+  babel: {
+    presets: [],
+    plugins: [
+      require('babel-plugin-module-resolver', {
+        "root": ["."],
+        "extensions": [".js", ".jsx"],
+        "alias": {
+          "@components": "./src/components"
+        }
+      })
+    ]
+  }
 }
