+++ src/components/ProductCard/index.jsx
@@ -1,11 +1,11 @@
-import React from 'react'
+import React from 'react';
 
-const ProductCard = () => {
+function ProductCard() {
   return (
     <div>
       <h1>Product Card</h1>
     </div>
-  )
+  );
 }
 
-export default ProductCard
+export default ProductCard;
